Add handleAddTactic function
